+++ lib/libesp32/berry_matter/src/solidify/solidified_Matter_Plugin_0.h
@@ -4,8 +4,8 @@
 \********************************************************************/
 #include "be_constobj.h"
 extern const bclass be_class_Matter_Plugin;
-// compact class 'Matter_Plugin' ktab size: 65, total: 108 (saved 344 bytes)
-static const bvalue be_ktab_class_Matter_Plugin[65] = {
+// compact class 'Matter_Plugin' ktab size: 66, total: 108 (saved 336 bytes)
+static const bvalue be_ktab_class_Matter_Plugin[66] = {
   /* K0   */  be_nested_str_weak(json),
   /* K1   */  be_nested_str_weak(node_label),
   /* K2   */  be_nested_str_weak(_X2C_X22Name_X22_X3A_X25s),
@@ -21,7 +21,7 @@ static const bvalue be_ktab_class_Matter_Plugin[65] = {
   /* K12  */  be_nested_str_weak(message_handler),
   /* K13  */  be_nested_str_weak(im),
   /* K14  */  be_nested_str_weak(send_ack_now),
-  /* K15  */  be_nested_str_weak(CLUSTERS),
+  /* K15  */  be_nested_str_weak(get_clusters),
   /* K16  */  be_nested_str_weak(find),
   /* K17  */  be_nested_str_weak(attribute_updated),
   /* K18  */  be_nested_str_weak(update_next),
@@ -53,24 +53,25 @@ static const bvalue be_ktab_class_Matter_Plugin[65] = {
   /* K44  */  be_nested_str_weak(_X25s_X2C_X25s_X3A_X25s),
   /* K45  */  be_nested_str_weak(publish_command),
   /* K46  */  be_nested_str_weak(MtrReceived),
-  /* K47  */  be_nested_str_weak(contains),
-  /* K48  */  be_nested_str_weak(TLV),
-  /* K49  */  be_nested_str_weak(cluster),
-  /* K50  */  be_nested_str_weak(attribute),
-  /* K51  */  be_nested_str_weak(Matter_TLV_array),
-  /* K52  */  be_nested_str_weak(TYPES),
-  /* K53  */  be_nested_str_weak(keys),
-  /* K54  */  be_nested_str_weak(add_struct),
-  /* K55  */  be_nested_str_weak(add_TLV),
-  /* K56  */  be_nested_str_weak(U2),
-  /* K57  */  be_nested_str_weak(stop_iteration),
-  /* K58  */  be_nested_str_weak(get_cluster_list_sorted),
-  /* K59  */  be_nested_str_weak(U4),
-  /* K60  */  be_const_int(3),
-  /* K61  */  be_nested_str_weak(set),
-  /* K62  */  be_nested_str_weak(get_attribute_list_bytes),
-  /* K63  */  be_nested_str_weak(FEATURE_MAPS),
-  /* K64  */  be_nested_str_weak(CLUSTER_REVISIONS),
+  /* K47  */  be_nested_str_weak(CLUSTERS),
+  /* K48  */  be_nested_str_weak(contains),
+  /* K49  */  be_nested_str_weak(TLV),
+  /* K50  */  be_nested_str_weak(cluster),
+  /* K51  */  be_nested_str_weak(attribute),
+  /* K52  */  be_nested_str_weak(Matter_TLV_array),
+  /* K53  */  be_nested_str_weak(TYPES),
+  /* K54  */  be_nested_str_weak(keys),
+  /* K55  */  be_nested_str_weak(add_struct),
+  /* K56  */  be_nested_str_weak(add_TLV),
+  /* K57  */  be_nested_str_weak(U2),
+  /* K58  */  be_nested_str_weak(stop_iteration),
+  /* K59  */  be_nested_str_weak(get_cluster_list_sorted),
+  /* K60  */  be_nested_str_weak(U4),
+  /* K61  */  be_const_int(3),
+  /* K62  */  be_nested_str_weak(set),
+  /* K63  */  be_nested_str_weak(get_attribute_list_bytes),
+  /* K64  */  be_nested_str_weak(FEATURE_MAPS),
+  /* K65  */  be_nested_str_weak(CLUSTER_REVISIONS),
 };
 
 
@@ -210,13 +211,14 @@ be_local_closure(class_Matter_Plugin_get_attribute_list_bytes,   /* name */
     &be_ktab_class_Matter_Plugin,     /* shared constants */
     be_str_weak(get_attribute_list_bytes),
     &be_const_str_solidified,
-    ( &(const binstruction[ 6]) {  /* code */
-      0x8808010F,  //  0000  GETMBR	R2	R0	K15
-      0x8C080510,  //  0001  GETMET	R2	R2	K16
-      0x5C100200,  //  0002  MOVE	R4	R1
-      0x4C140000,  //  0003  LDNIL	R5
-      0x7C080600,  //  0004  CALL	R2	3
-      0x80040400,  //  0005  RET	1	R2
+    ( &(const binstruction[ 7]) {  /* code */
+      0x8C08010F,  //  0000  GETMET	R2	R0	K15
+      0x7C080200,  //  0001  CALL	R2	1
+      0x8C080510,  //  0002  GETMET	R2	R2	K16
+      0x5C100200,  //  0003  MOVE	R4	R1
+      0x4C140000,  //  0004  LDNIL	R5
+      0x7C080600,  //  0005  CALL	R2	3
+      0x80040400,  //  0006  RET	1	R2
     })
   )
 );
@@ -612,7 +614,7 @@ be_local_closure(class_Matter_Plugin_consolidate_update_commands,   /* name */
 ********************************************************************/
 be_local_closure(class_Matter_Plugin_get_cluster_list_sorted,   /* name */
   be_nested_proto(
-    4,                          /* nstack */
+    5,                          /* nstack */
     1,                          /* argc */
     10,                          /* varg */
     0,                          /* has upvals */
@@ -623,12 +625,13 @@ be_local_closure(class_Matter_Plugin_get_cluster_list_sorted,   /* name */
     &be_ktab_class_Matter_Plugin,     /* shared constants */
     be_str_weak(get_cluster_list_sorted),
     &be_const_str_solidified,
-    ( &(const binstruction[ 5]) {  /* code */
+    ( &(const binstruction[ 6]) {  /* code */
       0x88040109,  //  0000  GETMBR	R1	R0	K9
       0x8C040320,  //  0001  GETMET	R1	R1	K32
-      0x880C010F,  //  0002  GETMBR	R3	R0	K15
-      0x7C040400,  //  0003  CALL	R1	2
-      0x80040200,  //  0004  RET	1	R1
+      0x8C0C010F,  //  0002  GETMET	R3	R0	K15
+      0x7C0C0200,  //  0003  CALL	R3	1
+      0x7C040400,  //  0004  CALL	R1	2
+      0x80040200,  //  0005  RET	1	R1
     })
   )
 );
@@ -820,33 +823,34 @@ be_local_closure(class_Matter_Plugin_contains_attribute,   /* name */
     &be_ktab_class_Matter_Plugin,     /* shared constants */
     be_str_weak(contains_attribute),
     &be_const_str_solidified,
-    ( &(const binstruction[26]) {  /* code */
-      0x880C010F,  //  0000  GETMBR	R3	R0	K15
-      0x8C0C0710,  //  0001  GETMET	R3	R3	K16
-      0x5C140200,  //  0002  MOVE	R5	R1
-      0x7C0C0400,  //  0003  CALL	R3	2
-      0x4C100000,  //  0004  LDNIL	R4
-      0x20100604,  //  0005  NE	R4	R3	R4
-      0x78120010,  //  0006  JMPF	R4	#0018
-      0x58100023,  //  0007  LDCONST	R4	K35
-      0x6014000C,  //  0008  GETGBL	R5	G12
-      0x5C180600,  //  0009  MOVE	R6	R3
-      0x7C140200,  //  000A  CALL	R5	1
-      0x0C140B24,  //  000B  DIV	R5	R5	K36
-      0x14180805,  //  000C  LT	R6	R4	R5
-      0x781A0009,  //  000D  JMPF	R6	#0018
-      0x8C180725,  //  000E  GETMET	R6	R3	K37
-      0x08200924,  //  000F  MUL	R8	R4	K36
-      0x5425FFFD,  //  0010  LDINT	R9	-2
-      0x7C180600,  //  0011  CALL	R6	3
-      0x1C180C02,  //  0012  EQ	R6	R6	R2
-      0x781A0001,  //  0013  JMPF	R6	#0016
-      0x50180200,  //  0014  LDBOOL	R6	1	0
-      0x80040C00,  //  0015  RET	1	R6
-      0x00100926,  //  0016  ADD	R4	R4	K38
-      0x7001FFF3,  //  0017  JMP		#000C
-      0x50100000,  //  0018  LDBOOL	R4	0	0
-      0x80040800,  //  0019  RET	1	R4
+    ( &(const binstruction[27]) {  /* code */
+      0x8C0C010F,  //  0000  GETMET	R3	R0	K15
+      0x7C0C0200,  //  0001  CALL	R3	1
+      0x8C0C0710,  //  0002  GETMET	R3	R3	K16
+      0x5C140200,  //  0003  MOVE	R5	R1
+      0x7C0C0400,  //  0004  CALL	R3	2
+      0x4C100000,  //  0005  LDNIL	R4
+      0x20100604,  //  0006  NE	R4	R3	R4
+      0x78120010,  //  0007  JMPF	R4	#0019
+      0x58100023,  //  0008  LDCONST	R4	K35
+      0x6014000C,  //  0009  GETGBL	R5	G12
+      0x5C180600,  //  000A  MOVE	R6	R3
+      0x7C140200,  //  000B  CALL	R5	1
+      0x0C140B24,  //  000C  DIV	R5	R5	K36
+      0x14180805,  //  000D  LT	R6	R4	R5
+      0x781A0009,  //  000E  JMPF	R6	#0019
+      0x8C180725,  //  000F  GETMET	R6	R3	K37
+      0x08200924,  //  0010  MUL	R8	R4	K36
+      0x5425FFFD,  //  0011  LDINT	R9	-2
+      0x7C180600,  //  0012  CALL	R6	3
+      0x1C180C02,  //  0013  EQ	R6	R6	R2
+      0x781A0001,  //  0014  JMPF	R6	#0017
+      0x50180200,  //  0015  LDBOOL	R6	1	0
+      0x80040C00,  //  0016  RET	1	R6
+      0x00100926,  //  0017  ADD	R4	R4	K38
+      0x7001FFF3,  //  0018  JMP		#000D
+      0x50100000,  //  0019  LDBOOL	R4	0	0
+      0x80040800,  //  001A  RET	1	R4
     })
   )
 );
@@ -1086,7 +1090,7 @@ be_local_closure(class_Matter_Plugin_get_clusters,   /* name */
     be_str_weak(get_clusters),
     &be_const_str_solidified,
     ( &(const binstruction[ 2]) {  /* code */
-      0x8804010F,  //  0000  GETMBR	R1	R0	K15
+      0x8804012F,  //  0000  GETMBR	R1	R0	K47
       0x80040200,  //  0001  RET	1	R1
     })
   )
@@ -1110,12 +1114,13 @@ be_local_closure(class_Matter_Plugin_contains_cluster,   /* name */
     &be_ktab_class_Matter_Plugin,     /* shared constants */
     be_str_weak(contains_cluster),
     &be_const_str_solidified,
-    ( &(const binstruction[ 5]) {  /* code */
-      0x8808010F,  //  0000  GETMBR	R2	R0	K15
-      0x8C08052F,  //  0001  GETMET	R2	R2	K47
-      0x5C100200,  //  0002  MOVE	R4	R1
-      0x7C080400,  //  0003  CALL	R2	2
-      0x80040400,  //  0004  RET	1	R2
+    ( &(const binstruction[ 6]) {  /* code */
+      0x8C08010F,  //  0000  GETMET	R2	R0	K15
+      0x7C080200,  //  0001  CALL	R2	1
+      0x8C080530,  //  0002  GETMET	R2	R2	K48
+      0x5C100200,  //  0003  MOVE	R4	R1
+      0x7C080400,  //  0004  CALL	R2	2
+      0x80040400,  //  0005  RET	1	R2
     })
   )
 );
@@ -1140,81 +1145,81 @@ be_local_closure(class_Matter_Plugin_read_attribute,   /* name */
     &be_const_str_solidified,
     ( &(const binstruction[169]) {  /* code */
       0xB8122600,  //  0000  GETNGBL	R4	K19
-      0x88100930,  //  0001  GETMBR	R4	R4	K48
-      0x88140531,  //  0002  GETMBR	R5	R2	K49
-      0x88180532,  //  0003  GETMBR	R6	R2	K50
+      0x88100931,  //  0001  GETMBR	R4	R4	K49
+      0x88140532,  //  0002  GETMBR	R5	R2	K50
+      0x88180533,  //  0003  GETMBR	R6	R2	K51
       0x541E001C,  //  0004  LDINT	R7	29
       0x1C1C0A07,  //  0005  EQ	R7	R5	R7
       0x781E0050,  //  0006  JMPF	R7	#0058
       0x1C1C0D23,  //  0007  EQ	R7	R6	K35
       0x781E001B,  //  0008  JMPF	R7	#0025
-      0x8C1C0933,  //  0009  GETMET	R7	R4	K51
+      0x8C1C0934,  //  0009  GETMET	R7	R4	K52
       0x7C1C0200,  //  000A  CALL	R7	1
-      0x88200134,  //  000B  GETMBR	R8	R0	K52
+      0x88200135,  //  000B  GETMBR	R8	R0	K53
       0x60240010,  //  000C  GETGBL	R9	G16
-      0x8C281135,  //  000D  GETMET	R10	R8	K53
+      0x8C281136,  //  000D  GETMET	R10	R8	K54
       0x7C280200,  //  000E  CALL	R10	1
       0x7C240200,  //  000F  CALL	R9	1
       0xA802000E,  //  0010  EXBLK	0	#0020
       0x5C281200,  //  0011  MOVE	R10	R9
       0x7C280000,  //  0012  CALL	R10	0
-      0x8C2C0F36,  //  0013  GETMET	R11	R7	K54
+      0x8C2C0F37,  //  0013  GETMET	R11	R7	K55
       0x7C2C0200,  //  0014  CALL	R11	1
-      0x8C301737,  //  0015  GETMET	R12	R11	K55
+      0x8C301738,  //  0015  GETMET	R12	R11	K56
       0x58380023,  //  0016  LDCONST	R14	K35
-      0x883C0938,  //  0017  GETMBR	R15	R4	K56
+      0x883C0939,  //  0017  GETMBR	R15	R4	K57
       0x5C401400,  //  0018  MOVE	R16	R10
       0x7C300800,  //  0019  CALL	R12	4
-      0x8C301737,  //  001A  GETMET	R12	R11	K55
+      0x8C301738,  //  001A  GETMET	R12	R11	K56
       0x58380026,  //  001B  LDCONST	R14	K38
-      0x883C0938,  //  001C  GETMBR	R15	R4	K56
+      0x883C0939,  //  001C  GETMBR	R15	R4	K57
       0x9440100A,  //  001D  GETIDX	R16	R8	R10
       0x7C300800,  //  001E  CALL	R12	4
       0x7001FFF0,  //  001F  JMP		#0011
-      0x58240039,  //  0020  LDCONST	R9	K57
+      0x5824003A,  //  0020  LDCONST	R9	K58
       0xAC240200,  //  0021  CATCH	R9	1	0
       0xB0080000,  //  0022  RAISE	2	R0	R0
       0x80040E00,  //  0023  RET	1	R7
       0x70020032,  //  0024  JMP		#0058
       0x1C1C0D26,  //  0025  EQ	R7	R6	K38
       0x781E0013,  //  0026  JMPF	R7	#003B
-      0x8C1C0933,  //  0027  GETMET	R7	R4	K51
+      0x8C1C0934,  //  0027  GETMET	R7	R4	K52
       0x7C1C0200,  //  0028  CALL	R7	1
       0x60200010,  //  0029  GETGBL	R8	G16
-      0x8C24013A,  //  002A  GETMET	R9	R0	K58
+      0x8C24013B,  //  002A  GETMET	R9	R0	K59
       0x7C240200,  //  002B  CALL	R9	1
       0x7C200200,  //  002C  CALL	R8	1
       0xA8020007,  //  002D  EXBLK	0	#0036
       0x5C241000,  //  002E  MOVE	R9	R8
       0x7C240000,  //  002F  CALL	R9	0
-      0x8C280F37,  //  0030  GETMET	R10	R7	K55
+      0x8C280F38,  //  0030  GETMET	R10	R7	K56
       0x4C300000,  //  0031  LDNIL	R12
-      0x8834093B,  //  0032  GETMBR	R13	R4	K59
+      0x8834093C,  //  0032  GETMBR	R13	R4	K60
       0x5C381200,  //  0033  MOVE	R14	R9
       0x7C280800,  //  0034  CALL	R10	4
       0x7001FFF7,  //  0035  JMP		#002E
-      0x58200039,  //  0036  LDCONST	R8	K57
+      0x5820003A,  //  0036  LDCONST	R8	K58
       0xAC200200,  //  0037  CATCH	R8	1	0
       0xB0080000,  //  0038  RAISE	2	R0	R0
       0x80040E00,  //  0039  RET	1	R7
       0x7002001C,  //  003A  JMP		#0058
       0x1C1C0D24,  //  003B  EQ	R7	R6	K36
       0x781E0003,  //  003C  JMPF	R7	#0041
-      0x8C1C0933,  //  003D  GETMET	R7	R4	K51
+      0x8C1C0934,  //  003D  GETMET	R7	R4	K52
       0x7C1C0200,  //  003E  CALL	R7	1
       0x80040E00,  //  003F  RET	1	R7
       0x70020016,  //  0040  JMP		#0058
-      0x1C1C0D3C,  //  0041  EQ	R7	R6	K60
+      0x1C1C0D3D,  //  0041  EQ	R7	R6	K61
       0x781E0003,  //  0042  JMPF	R7	#0047
-      0x8C1C0933,  //  0043  GETMET	R7	R4	K51
+      0x8C1C0934,  //  0043  GETMET	R7	R4	K52
       0x7C1C0200,  //  0044  CALL	R7	1
       0x80040E00,  //  0045  RET	1	R7
       0x70020010,  //  0046  JMP		#0058
       0x541EFFFB,  //  0047  LDINT	R7	65532
       0x1C1C0C07,  //  0048  EQ	R7	R6	R7
       0x781E0005,  //  0049  JMPF	R7	#0050
-      0x8C1C073D,  //  004A  GETMET	R7	R3	K61
-      0x8824093B,  //  004B  GETMBR	R9	R4	K59
+      0x8C1C073E,  //  004A  GETMET	R7	R3	K62
+      0x8824093C,  //  004B  GETMBR	R9	R4	K60
       0x58280023,  //  004C  LDCONST	R10	K35
       0x7C1C0600,  //  004D  CALL	R7	3
       0x80040E00,  //  004E  RET	1	R7
@@ -1222,24 +1227,24 @@ be_local_closure(class_Matter_Plugin_read_attribute,   /* name */
       0x541EFFFC,  //  0050  LDINT	R7	65533
       0x1C1C0C07,  //  0051  EQ	R7	R6	R7
       0x781E0004,  //  0052  JMPF	R7	#0058
-      0x8C1C073D,  //  0053  GETMET	R7	R3	K61
-      0x8824093B,  //  0054  GETMBR	R9	R4	K59
+      0x8C1C073E,  //  0053  GETMET	R7	R3	K62
+      0x8824093C,  //  0054  GETMBR	R9	R4	K60
       0x58280026,  //  0055  LDCONST	R10	K38
       0x7C1C0600,  //  0056  CALL	R7	3
       0x80040E00,  //  0057  RET	1	R7
       0x541EFFF7,  //  0058  LDINT	R7	65528
       0x1C1C0C07,  //  0059  EQ	R7	R6	R7
       0x781E0003,  //  005A  JMPF	R7	#005F
-      0x8C1C0933,  //  005B  GETMET	R7	R4	K51
+      0x8C1C0934,  //  005B  GETMET	R7	R4	K52
       0x7C1C0200,  //  005C  CALL	R7	1
       0x80040E00,  //  005D  RET	1	R7
       0x70020047,  //  005E  JMP		#00A7
       0x541EFFFA,  //  005F  LDINT	R7	65531
       0x1C1C0C07,  //  0060  EQ	R7	R6	R7
       0x781E001B,  //  0061  JMPF	R7	#007E
-      0x8C1C0933,  //  0062  GETMET	R7	R4	K51
+      0x8C1C0934,  //  0062  GETMET	R7	R4	K52
       0x7C1C0200,  //  0063  CALL	R7	1
-      0x8C20013E,  //  0064  GETMET	R8	R0	K62
+      0x8C20013F,  //  0064  GETMET	R8	R0	K63
       0x5C280A00,  //  0065  MOVE	R10	R5
       0x7C200400,  //  0066  CALL	R8	2
       0x4C240000,  //  0067  LDNIL	R9
@@ -1253,9 +1258,9 @@ be_local_closure(class_Matter_Plugin_read_attribute,   /* name */
       0x58280023,  //  006F  LDCONST	R10	K35
       0x142C1409,  //  0070  LT	R11	R10	R9
       0x782E0009,  //  0071  JMPF	R11	#007C
-      0x8C2C0F37,  //  0072  GETMET	R11	R7	K55
+      0x8C2C0F38,  //  0072  GETMET	R11	R7	K56
       0x4C340000,  //  0073  LDNIL	R13
-      0x88380938,  //  0074  GETMBR	R14	R4	K56
+      0x88380939,  //  0074  GETMBR	R14	R4	K57
       0x8C3C1125,  //  0075  GETMET	R15	R8	K37
       0x08441524,  //  0076  MUL	R17	R10	K36
       0x5449FFFD,  //  0077  LDINT	R18	-2
@@ -1268,27 +1273,27 @@ be_local_closure(class_Matter_Plugin_read_attribute,   /* name */
       0x541EFFF9,  //  007E  LDINT	R7	65530
       0x1C1C0C07,  //  007F  EQ	R7	R6	R7
       0x781E0003,  //  0080  JMPF	R7	#0085
-      0x8C1C0933,  //  0081  GETMET	R7	R4	K51
+      0x8C1C0934,  //  0081  GETMET	R7	R4	K52
       0x7C1C0200,  //  0082  CALL	R7	1
       0x80040E00,  //  0083  RET	1	R7
       0x70020021,  //  0084  JMP		#00A7
       0x541EFFF8,  //  0085  LDINT	R7	65529
       0x1C1C0C07,  //  0086  EQ	R7	R6	R7
       0x781E0003,  //  0087  JMPF	R7	#008C
-      0x8C1C0933,  //  0088  GETMET	R7	R4	K51
+      0x8C1C0934,  //  0088  GETMET	R7	R4	K52
       0x7C1C0200,  //  0089  CALL	R7	1
       0x80040E00,  //  008A  RET	1	R7
       0x7002001A,  //  008B  JMP		#00A7
       0x541EFFFB,  //  008C  LDINT	R7	65532
       0x1C1C0C07,  //  008D  EQ	R7	R6	R7
       0x781E000A,  //  008E  JMPF	R7	#009A
-      0x881C013F,  //  008F  GETMBR	R7	R0	K63
+      0x881C0140,  //  008F  GETMBR	R7	R0	K64
       0x8C1C0F10,  //  0090  GETMET	R7	R7	K16
       0x5C240A00,  //  0091  MOVE	R9	R5
       0x58280023,  //  0092  LDCONST	R10	K35
       0x7C1C0600,  //  0093  CALL	R7	3
-      0x8C20073D,  //  0094  GETMET	R8	R3	K61
-      0x8828093B,  //  0095  GETMBR	R10	R4	K59
+      0x8C20073E,  //  0094  GETMET	R8	R3	K62
+      0x8828093C,  //  0095  GETMBR	R10	R4	K60
       0x5C2C0E00,  //  0096  MOVE	R11	R7
       0x7C200600,  //  0097  CALL	R8	3
       0x80041000,  //  0098  RET	1	R8
@@ -1296,13 +1301,13 @@ be_local_closure(class_Matter_Plugin_read_attribute,   /* name */
       0x541EFFFC,  //  009A  LDINT	R7	65533
       0x1C1C0C07,  //  009B  EQ	R7	R6	R7
       0x781E0009,  //  009C  JMPF	R7	#00A7
-      0x881C0140,  //  009D  GETMBR	R7	R0	K64
+      0x881C0141,  //  009D  GETMBR	R7	R0	K65
       0x8C1C0F10,  //  009E  GETMET	R7	R7	K16
       0x5C240A00,  //  009F  MOVE	R9	R5
       0x58280026,  //  00A0  LDCONST	R10	K38
       0x7C1C0600,  //  00A1  CALL	R7	3
-      0x8C20073D,  //  00A2  GETMET	R8	R3	K61
-      0x8828093B,  //  00A3  GETMBR	R10	R4	K59
+      0x8C20073E,  //  00A2  GETMET	R8	R3	K62
+      0x8828093C,  //  00A3  GETMBR	R10	R4	K60
       0x5C2C0E00,  //  00A4  MOVE	R11	R7
       0x7C200600,  //  00A5  CALL	R8	3
       0x80041000,  //  00A6  RET	1	R8